+++ app/src/main/java/np/com/naxa/drone_tasking_manager/features/user/task_dashboard/models/UsersTaskItem.kt
@@ -4,7 +4,7 @@ data class UsersTaskItem(
     val certificateUrl: String?,
     val createdAt: String?,
     val flightDistanceKm: Double?,
-    val flightTimeMinutes: Int?,
+    val flightTimeMinutes: Double?,
     val projectId: String?,
     val projectName: String?,
     val projectTaskIndex: Int?,
